+++ lib/benchee_dsl/runner.ex
@@ -130,23 +130,36 @@ defmodule BencheeDsl.Runner do
   defp jobs(module, %{jobs: jobs}) do
     Enum.reduce(jobs, %{}, fn {job, opts}, acc ->
       tags = Keyword.get(opts, :tags)
-      before = opts |> Keyword.get(:before) |> before()
       name = to_string(job)
       fun = module.job(job)
+      job_opts = job_opts(opts)
 
-      case {Enum.member?(tags, :skip), before} do
+      case {Enum.member?(tags, :skip), job_opts} do
         {true, _} -> acc
-        {false, nil} -> Map.put(acc, name, fun)
-        {false, before} -> Map.put(acc, name, {fun, before_scenario: before})
+        {false, []} -> Map.put(acc, name, fun)
+        {false, job_opts} -> Map.put(acc, name, {fun, job_opts})
       end
     end)
   end
 
   defp jobs(_, _), do: %{}
 
-  defp before(nil), do: nil
+  defp job_opts(opts) do
+    []
+    |> hooks(opts, :setup, :before_scenario)
+    |> hooks(opts, :setup_each, :before_each)
+    |> hooks(opts, :on_exit, :after_scenario)
+    |> hooks(opts, :on_exit_each, :after_each)
+  end
 
-  defp before(fun) do
+  defp hooks(hooks, opts, tag, key) do
+    case Keyword.get(opts, tag) do
+      nil -> hooks
+      fun -> Keyword.put(hooks, key, hook(fun))
+    end
+  end
+
+  defp hook(fun) do
     case Function.info(fun, :arity) do
       {:arity, 0} ->
         fn x ->