+++ test/benchee_dsl_test.exs
@@ -318,24 +318,35 @@ defmodule BencheeDslTest do
     assert_run("test/fixtures/jobs_update_bench.exs")
   end
 
-  test "runs benchmark for before_bench.exs" do
+  test "runs benchmark for hooks_bench.exs" do
     expect(BencheeMock, :run, fn jobs, config ->
       assert %{
-               "flat_map" => {flat_map, before_scenario: flat_map_before},
-               "map.flatten" => {map_flatten, before_scenario: map_flatten_before}
+               "flat_map" => {flat_map, flat_map_hooks},
+               "map_flatten" => {map_flatten, map_flatten_hooks}
              } = jobs
 
       assert is_function(flat_map, 1)
-      assert is_function(flat_map_before, 1)
+      assert is_function(flat_map_hooks[:before_scenario], 1)
+      assert flat_map_hooks[:before_scenario].([1,2]) == 1..2
+      assert is_function(flat_map_hooks[:before_each], 1)
+      assert flat_map_hooks[:before_each].(:in) == :in
+      assert is_function(flat_map_hooks[:after_scenario], 1)
+      assert is_function(flat_map_hooks[:after_each], 1)
+
       assert is_function(map_flatten, 1)
-      assert is_function(map_flatten_before, 1)
-      assert map_flatten_before.(:in) == :in
+      assert is_function(map_flatten_hooks[:before_scenario], 1)
+      assert map_flatten_hooks[:before_scenario].(:in) == :in
+      assert is_function(map_flatten_hooks[:before_each], 1)
+      assert map_flatten_hooks[:before_each].(:in) == :in
+      assert is_function(map_flatten_hooks[:after_scenario], 1)
+      assert is_function(map_flatten_hooks[:after_each], 1)
 
       benchee_run(jobs, config)
     end)
 
-    assert_run("test/fixtures/before_bench.exs")
+    assert_run("test/fixtures/hooks_bench.exs")
   end
+
 
   defp assert_run(file) do
     assert BencheeDsl.config(file: file)
@@ -357,8 +368,6 @@ defmodule BencheeDslTest do
   end
 
   defp benchee_run(jobs, config) do
-    if @benchee_run do
-      assert Benchee.run(jobs, config)
-    end
+    if @benchee_run, do: assert Benchee.run(jobs, config)
   end
 end