+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0 - dev
+
++ Add local hooks (`@setup`, `@setup_each`, `@on_exit`, `@on_exit_each`)
+
+### Breaking changes
+
++ Replace `@before` with `@setup`.
+
 ## 0.1.4 2022/07/16
 
 + Add `recode` dependency.